requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
httpx==0.28.1
Werkzeug==3.0.1
Jinja2==3.1.2